applied api key and controller processes json

+++ WeatherAPI/Class/Main.cs
@@ -2,7 +2,7 @@
 {
     public class Main
     {
-        public double Temp { get; set; }
+        public float Temp { get; set; }
         public int Pressure { get; set; }
         public int Humidity { get; set; }
         public int Temp_Min { get; set; }
